Safari-609.1.11.1/LayoutTests/accessibility/mac/aria-current-attribute-exposed-expected.txt

Link Current Link
This tests that aria-current attribute is correctly exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS link1.allAttributes().includes('AXARIACurrent: true') is false
PASS link2.allAttributes().includes('AXARIACurrent: true') is true
PASS successfullyParsed is true

TEST COMPLETE
